+++ dds_cli/file_handler.py
@@ -49,10 +49,9 @@ class FileHandler:
                     with source_path_file.resolve().open(mode="r") as spf:
                         self.data_list += spf.read().splitlines()
                 except OSError as err:
-                    dds_cli.utils.console.print(
+                    raise dds_cli.exceptions.UploadError(
                         f"Failed to get files from source-path-file option: {err}"
                     )
-                    os._exit(1)
 
         self.failed = {}